ULTIMATE_v193_COMPLETE/requirements.txt
# ============================================================================
# UNIFIED TRADING SYSTEM v1.3.15.90 - CENTRAL DEPENDENCIES
# ============================================================================
# Single dependency file for ALL components:
#   - FinBERT v4.4.4 (sentiment analysis + LSTM training)
#   - Ultimate Trading Dashboard (swing trading + paper trading)
#   - Pipelines (AU/US/UK overnight screening)
#
# Python 3.12+ compatible
# Pre-built wheels for Windows 11 (no build tools required)
# Tested: 2026-02-05
# ============================================================================

# ============================================
# Core Framework (Flask + Dashboard)
# ============================================
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.0
python-dotenv==1.0.0
dash==2.14.2
plotly==5.18.0
dash-bootstrap-components==1.5.0

# ============================================
# Data Processing (ALL components)
# ============================================
# Using pandas 2.2.0+ for pre-built Windows wheels (no Visual Studio needed)
numpy>=1.26.0
pandas>=2.2.0
python-dateutil==2.8.2
pytz>=2023.3

# ============================================
# Market Data (ALL components)
# ============================================
yfinance>=0.2.28
yahooquery>=2.3.0

# ============================================
# Web Scraping (FinBERT + Pipelines)
# ============================================
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0
requests>=2.31.0
feedparser>=6.0.10

# ============================================
# Machine Learning - TensorFlow (LSTM + Dashboard)
# ============================================
tensorflow==2.16.1
# Note: Keras included in TensorFlow 2.16.1 (use tf.keras)

# ============================================
# Machine Learning - PyTorch (FinBERT Sentiment)
# SECURITY FIX: PyTorch 2.6.0 fixes CVE-2025-32434
# ============================================
torch>=2.6.0
torchvision>=0.21.0
transformers>=4.36.0
sentencepiece>=0.1.99

# ============================================
# ML Utilities (ALL components)
# ============================================
scikit-learn>=1.3.0
hmmlearn>=0.3.0

# ============================================
# AI & LLM (v192+v193)
# ============================================
google-generativeai>=0.3.0

# ============================================
# Technical Analysis (ALL components)
# ============================================
ta>=0.10.0

# ============================================
# API & Utilities (ALL components)
# ============================================
urllib3>=1.26.0,<2.0.0
websockets>=11.0.0
python-multipart>=0.0.9

# ============================================
# Optional: Performance
# ============================================
# tokenizers>=0.13.0  # Requires Rust compiler - uncomment if needed
